// Assets-Uncapped/front-portal-develop/src/services/api/connections/models/Bank.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-connections
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: Unknown version
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    Image,
    ImageFromJSON,
    ImageFromJSONTyped,
    ImageToJSON,
} from './Image';

/**
 * 
 * @export
 * @interface Bank
 */
export interface Bank {
    /**
     * 
     * @type {string}
     * @memberof Bank
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof Bank
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof Bank
     */
    country?: string;
    /**
     * 
     * @type {Image}
     * @memberof Bank
     */
    logo?: Image;
    /**
     * 
     * @type {Image}
     * @memberof Bank
     */
    icon?: Image;
    /**
     * 
     * @type {string}
     * @memberof Bank
     */
    systemId?: string;
    /**
     * 
     * @type {string}
     * @memberof Bank
     */
    logoUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof Bank
     */
    iconUrl?: string;
}

export function BankFromJSON(json: any): Bank {
    return BankFromJSONTyped(json, false);
}

export function BankFromJSONTyped(json: any, ignoreDiscriminator: boolean): Bank {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'country': !exists(json, 'country') ? undefined : json['country'],
        'logo': !exists(json, 'logo') ? undefined : ImageFromJSON(json['logo']),
        'icon': !exists(json, 'icon') ? undefined : ImageFromJSON(json['icon']),
        'systemId': !exists(json, 'systemId') ? undefined : json['systemId'],
        'logoUrl': !exists(json, 'logoUrl') ? undefined : json['logoUrl'],
        'iconUrl': !exists(json, 'iconUrl') ? undefined : json['iconUrl'],
    };
}

export function BankToJSON(value?: Bank | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'country': value.country,
        'logo': ImageToJSON(value.logo),
        'icon': ImageToJSON(value.icon),
        'systemId': value.systemId,
        'logoUrl': value.logoUrl,
        'iconUrl': value.iconUrl,
    };
}